backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-generativeai==0.3.2
requests==2.31.0
Pillow==10.0.1
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
gunicorn==21.2.0
